pyproject.toml
[tool.poetry]
name = "shop"
version = "0.1.0"
description = ""
authors = ["Vladislav"]
readme = "README.md"
packages = [{include = "shop", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12.0"
psycopg2 = "2.9.10"
djangorestframework = "3.15.2"
gunicorn = "23.0.0"
python-decouple = "3.8"
drf-spectacular = "0.28.0"
drf-spectacular-sidecar = "2024.12.1"
pillow = "^11.0.0"
factory-boy = "^3.3.1"
django-filter = "^24.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
